Complied entire book
While there is still much to do, I compiled the entire book just to see where I was on the page count. I have 104 references and 295 pages. The glossary did not print for some reason. There are a lot of other changes I need to make (adding graphics, for example). But it's a start...